+++ decidim-forms/lib/decidim/forms/test/shared_examples/manage_questionnaires/add_questions.rb
@@ -11,6 +11,8 @@ shared_examples_for "add questions" do
 
       expect(page).to have_selector(".questionnaire-question", count: 2)
 
+      expand_all_questions
+
       page.all(".questionnaire-question").each_with_index do |question, idx|
         within question do
           fill_in find_nested_form_field_locator("body_en"), with: questions_body[idx]
@@ -22,7 +24,7 @@ shared_examples_for "add questions" do
 
     expect(page).to have_admin_callout("successfully")
 
-    visit questionnaire_edit_path
+    visit_questionnaire_edit_path_and_expand_all
 
     expect(page).to have_selector("input[value='This is the first question']")
     expect(page).to have_selector("input[value='This is the second question']")
@@ -31,6 +33,7 @@ shared_examples_for "add questions" do
   it "adds a question with a rich text description" do
     within "form.edit_questionnaire" do
       click_button "Add question"
+      expand_all_questions
 
       within ".questionnaire-question" do
         fill_in find_nested_form_field_locator("body_en"), with: "Body"
@@ -74,6 +77,7 @@ shared_examples_for "add questions" do
     within "form.edit_questionnaire" do
       click_button "Add question"
       click_button "Add question"
+      expand_all_questions
 
       page.all(".questionnaire-question").each_with_index do |question, idx|
         within question do
@@ -103,7 +107,7 @@ shared_examples_for "add questions" do
 
     expect(page).to have_admin_callout("successfully")
 
-    visit questionnaire_edit_path
+    visit_questionnaire_edit_path_and_expand_all
 
     expect(page).to have_selector("input[value='This is the first question']")
     expect(page).to have_selector("input[value='This is the Q1 first option']")
@@ -117,6 +121,7 @@ shared_examples_for "add questions" do
 
   it "adds a sane number of options for each attribute type" do
     click_button "Add question"
+    expand_all_questions
 
     select "Long answer", from: "Type"
     expect(page).to have_no_selector(".questionnaire-question-answer-option")
@@ -145,6 +150,8 @@ shared_examples_for "add questions" do
 
   it "does not incorrectly reorder when clicking answer options" do
     click_button "Add question"
+    expand_all_questions
+
     select "Single option", from: "Type"
     2.times { click_button "Add answer option" }
 
@@ -173,6 +180,8 @@ shared_examples_for "add questions" do
 
   it "does not incorrectly reorder when clicking matrix rows" do
     click_button "Add question"
+    expand_all_questions
+
     select "Matrix (Multiple option)", from: "Type"
     2.times { click_button "Add row" }
 
@@ -201,14 +210,19 @@ shared_examples_for "add questions" do
 
   it "preserves question form across submission failures" do
     click_button "Add question"
+    expand_all_questions
+
     select "Long answer", from: "Type"
     click_button "Save"
 
+    expand_all_questions
     expect(page).to have_select("Type", selected: "Long answer")
   end
 
   it "does not preserve spurious answer options from previous type selections" do
     click_button "Add question"
+    expand_all_questions
+
     select "Single option", from: "Type"
 
     within ".questionnaire-question-answer-option:first-of-type" do
@@ -218,6 +232,7 @@ shared_examples_for "add questions" do
     select "Long answer", from: "Type"
 
     click_button "Save"
+    expand_all_questions
 
     select "Single option", from: "Type"
 
@@ -228,6 +243,8 @@ shared_examples_for "add questions" do
 
   it "does not preserve spurious matrix rows from previous type selections" do
     click_button "Add question"
+    expand_all_questions
+
     select "Matrix (Single option)", from: "Type"
 
     within ".questionnaire-question-matrix-row:first-of-type" do
@@ -237,6 +254,7 @@ shared_examples_for "add questions" do
     select "Long answer", from: "Type"
 
     click_button "Save"
+    expand_all_questions
 
     select "Matrix (Single option)", from: "Type"
 
@@ -247,6 +265,8 @@ shared_examples_for "add questions" do
 
   it "preserves answer options form across submission failures" do
     click_button "Add question"
+    expand_all_questions
+
     select "Multiple option", from: "Type"
 
     within ".questionnaire-question-answer-option:first-of-type" do
@@ -262,6 +282,7 @@ shared_examples_for "add questions" do
     select "3", from: "Maximum number of choices"
 
     click_button "Save"
+    expand_all_questions
 
     within ".questionnaire-question-answer-option:first-of-type" do
       expect(page).to have_nested_field("body_en", with: "Something")
@@ -276,6 +297,8 @@ shared_examples_for "add questions" do
 
   it "preserves matrix rows form across submission failures" do
     click_button "Add question"
+    expand_all_questions
+
     select "Matrix (Multiple option)", from: "Type"
 
     within ".questionnaire-question-matrix-row:first-of-type" do
@@ -285,6 +308,7 @@ shared_examples_for "add questions" do
     click_button "Add row"
 
     click_button "Save"
+    expand_all_questions
 
     within ".questionnaire-question-matrix-row:first-of-type" do
       expect(page).to have_nested_field("body_en", with: "Something")
@@ -294,6 +318,8 @@ shared_examples_for "add questions" do
   it "allows switching translated field tabs after form failures" do
     click_button "Add question"
     click_button "Save"
+
+    expand_all_questions
 
     within ".questionnaire-question:first-of-type" do
       fill_in find_nested_form_field_locator("body_en"), with: "Bye"
@@ -314,6 +340,8 @@ shared_examples_for "add questions" do
 
       within "form.edit_questionnaire" do
         click_button "Add question"
+
+        expand_all_questions
 
         within ".questionnaire-question" do
           fill_in find_nested_form_field_locator("body_en"), with: "This is the first question"
@@ -356,6 +384,7 @@ shared_examples_for "add questions" do
       expect(page).to have_select("Maximum number of choices", options: %w(Any 2))
 
       click_button "Add question"
+      expand_all_questions
 
       within(".questionnaire-question:last-of-type") do
         select "Multiple option", from: "Type"
@@ -373,6 +402,7 @@ shared_examples_for "add questions" do
 
       within "form.edit_questionnaire" do
         click_button "Add question"
+        expand_all_questions
 
         within ".questionnaire-question" do
           fill_in find_nested_form_field_locator("body_en"), with: "This is the first question"
@@ -416,6 +446,7 @@ shared_examples_for "add questions" do
       expect(page).to have_select("Maximum number of choices", options: %w(Any 2))
 
       click_button "Add question"
+      expand_all_questions
 
       within(".questionnaire-question:last-of-type") do
         select "Matrix (Multiple option)", from: "Type"